removed silly unsafes in scripting

+++ hyper_ast/src/scripting/mod.rs
@@ -42,5 +42,5 @@ impl<HAST, Acc> From<std::sync::Arc<str>> for Prepro<HAST, Acc> {
 }
 
 pub struct Acc {
-    id: usize,
+    id: u16,
 }
